test: extract shared harness helpers
